# Conflicts:
#	app/Models/Region.cs
#	app/app.csproj

+++ app/Models/Region.cs
@@ -5,11 +5,13 @@ using System.Web;
 
 namespace Marketing.Models
 {
+	/// <summary>
+	/// Заглушка для сущности Регион
+	/// </summary>
 	public class Region
 	{
-		public virtual uint Id { get; set; }
+		public virtual ulong Id { get; set; }
 
 		public virtual string Name { get; set; }
-
 	}
 }
